# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "F1 Race Analysis Engine",
    "description": "\nAdvanced Formula 1 analytics.\n\nEvery analysis response carries the `engine_version` that produced it. Cached results\nare keyed on that version, so changing a model invalidates its cached values rather\nthan mixing two definitions in one chart.\n\n**On estimates.** Degradation slopes, fuel-corrected lap times and pit loss are modelled\nfrom observed timing, not measured from team telemetry. Fields that carry an estimate\nsay so in their description. In particular the fuel coefficient is a published\n0.030 s/kg prior, not a value fitted from this dataset \u2014 it cannot be fitted from race\ndata, because fuel load is derived from lap number and carries no independent information.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/season/circuits": {
      "get": {
        "tags": [
          "season"
        ],
        "summary": "Circuit Profiles",
        "description": "Which circuits punish tyres, ranked.\n\nA strategist's first question about an unfamiliar track, and one that only has an\nanswer across a calendar: a single race tells you what happened there, not how the\ncircuit compares.",
        "operationId": "circuit_profiles_api_v1_season_circuits_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to one season",
              "title": "Season"
            },
            "description": "Restrict to one season"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/season/pace/{season}": {
      "get": {
        "tags": [
          "season"
        ],
        "summary": "Season Pace",
        "description": "Every driver's pace gap, race by race.\n\nThe shape of a season: who improved, who fell away, and where a car's upgrade\nactually landed. A per-race ranking cannot show any of that.",
        "operationId": "season_pace_api_v1_season_pace__season__get",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonPaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CircuitProfileOut": {
        "properties": {
          "circuit_key": {
            "type": "string",
            "title": "Circuit Key"
          },
          "circuit_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Circuit Name"
          },
          "n_races": {
            "type": "integer",
            "title": "N Races"
          },
          "seasons": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Seasons"
          },
          "degradation_s_per_lap": {
            "type": "number",
            "title": "Degradation S Per Lap",
            "description": "Median across compounds and races at this circuit"
          },
          "typical_stint_laps": {
            "type": "number",
            "title": "Typical Stint Laps",
            "description": "Longest stint observed, averaged \u2014 how long a set lasts here"
          },
          "net_pit_loss_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Pit Loss S",
            "description": "What a stop costs at this circuit"
          },
          "typical_stops": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typical Stops"
          },
          "reference_lap_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Lap S"
          }
        },
        "type": "object",
        "required": [
          "circuit_key",
          "n_races",
          "seasons",
          "degradation_s_per_lap",
          "typical_stint_laps"
        ],
        "title": "CircuitProfileOut",
        "description": "One circuit's tyre and strategy character, pooled across every race there."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Meta": {
        "properties": {
          "engine_version": {
            "type": "string",
            "title": "Engine Version",
            "description": "Engine version that produced these values"
          },
          "computed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Computed At",
            "description": "When the underlying metrics were computed"
          }
        },
        "type": "object",
        "required": [
          "engine_version"
        ],
        "title": "Meta",
        "description": "Provenance attached to every analysis response."
      },
...
      "SeasonPaceResponse": {
        "properties": {
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "rounds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Rounds"
          },
          "drivers": {
            "items": {
              "$ref": "#/components/schemas/SeasonPaceRowOut"
            },
            "type": "array",
            "title": "Drivers"
          }
        },
        "type": "object",
        "required": [
          "season",
          "meta",
          "rounds",
          "drivers"
        ],
        "title": "SeasonPaceResponse"
      },
      "SeasonPaceRowOut": {
        "properties": {
          "driver_number": {
            "type": "string",
            "title": "Driver Number"
          },
          "n_races": {
            "type": "integer",
            "title": "N Races"
          },
          "mean_gap_s": {
            "type": "number",
            "title": "Mean Gap S",
            "description": "Mean gap to the quickest car, per race"
          },
          "best_gap_s": {
            "type": "number",
            "title": "Best Gap S"
          },
          "worst_gap_s": {
            "type": "number",
            "title": "Worst Gap S"
          },
          "wins_on_pace": {
            "type": "integer",
            "title": "Wins On Pace",
            "description": "Races where this driver had the quickest corrected pace"
          },
          "gaps": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Gaps"
          }
        },
        "type": "object",
        "required": [
          "driver_number",
          "n_races",
          "mean_gap_s",
          "best_gap_s",
          "worst_gap_s",
          "wins_on_pace",
          "gaps"
        ],
        "title": "SeasonPaceRowOut",
        "description": "One driver's pace across a season, race by race."
      },
      "SeasonProfileResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season",
            "description": "Null when pooled across every ingested season"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": "Degradation is fitted from lap times, not measured. Circuits are ranked on the median across the races held there."
          },
          "circuits": {
            "items": {
              "$ref": "#/components/schemas/CircuitProfileOut"
            },
            "type": "array",
            "title": "Circuits"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "circuits"
        ],
        "title": "SeasonProfileResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}